+++ packages/react-ui/components/MenuItem/MenuItem.tsx
@@ -2,6 +2,7 @@ import React, { AriaAttributes, HTMLAttributes } from 'react';
 import PropTypes from 'prop-types';
 import { globalObject, isBrowser } from '@skbkontur/global-object';
 
+import { scrollYCenterIntoNearestScrollable } from '../../lib/dom/scrollYCenterIntoNearestScrollable';
 import { Nullable } from '../../typings/utility-types';
 import { isExternalLink, isFunction, isNonNullable, isReactUIComponent } from '../../lib/utils';
 import { ThemeContext } from '../../lib/theming/ThemeContext';
@@ -79,6 +80,7 @@ export interface MenuItemProps
 
   /** Устанавливает стиль для отображения в мобильной версии. */
   isMobile?: boolean;
+  /** @ignore */
   scrollIntoView?: boolean;
 }
 
@@ -156,10 +158,8 @@ export class MenuItem extends React.Component<MenuItemProps> {
   }
 
   public componentDidMount() {
-    if (this.props.scrollIntoView) {
-      this.rootRef?.scrollIntoView({
-        block: 'center',
-      });
+    if (this.props.scrollIntoView && this.rootRef) {
+      scrollYCenterIntoNearestScrollable(this.rootRef);
     }
     if (this.rootRef && isBrowser(globalObject)) {
       this.setState({ iconOffsetTop: globalObject.getComputedStyle(this.rootRef).getPropertyValue('padding-top') });