+++ packages/react-ui/lib/dom/getParentOrShadowHost.ts
@@ -1,0 +1,3 @@
+export function getParentOrShadowHost(element: Element) {
+  return element.parentElement || (element.getRootNode() as ShadowRoot).host;
+}